+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 
 * [HexaPDF::FontLoader::FromConfiguration] to accept arbitrary keyword arguments
 * [HexaPDF::Font::CMap::Parser] to avoid instantiating invalid UTF-16BE chars
+* [HexaPDF::Type::AcroForm::AppearanceGenerator] to work for files where check
+  boxes don't have appearance subdictionaries
 
 
 ## 0.40.0 - 2024-03-23